# Conflicts:
#	flickit-assessment-common/src/main/java/org/flickit/assessment/common/adapter/out/novu/NotificationType.java
#	flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/domain/AssessmentKit.java
#	flickit-assessment-kit/src/test/java/org/flickit/assessment/kit/test/fixture/application/AssessmentKitMother.java

+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/maturitylevel/MaturityLevelPersistenceJpaAdapter.java
@@ -1,24 +1,31 @@
 package org.flickit.assessment.kit.adapter.out.persistence.maturitylevel;
 
 import lombok.RequiredArgsConstructor;
-import org.flickit.assessment.data.jpa.kit.levelcompetence.LevelCompetenceJpaEntity;
+import org.flickit.assessment.common.application.domain.crud.PaginatedResponse;
+import org.flickit.assessment.common.exception.ResourceNotFoundException;
+import org.flickit.assessment.common.util.JsonUtils;
 import org.flickit.assessment.data.jpa.kit.levelcompetence.LevelCompetenceJpaRepository;
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaEntity;
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaEntity.EntityId;
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaRepository;
+import org.flickit.assessment.data.jpa.kit.seq.KitDbSequenceGenerators;
 import org.flickit.assessment.kit.application.domain.MaturityLevel;
-import org.flickit.assessment.kit.application.domain.MaturityLevelCompetence;
 import org.flickit.assessment.kit.application.port.out.maturitylevel.*;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import static java.util.stream.Collectors.toMap;
+import static org.flickit.assessment.kit.adapter.out.persistence.maturitylevel.MaturityLevelMapper.mapToDomainModel;
 import static org.flickit.assessment.kit.adapter.out.persistence.maturitylevel.MaturityLevelMapper.mapToJpaEntityToPersist;
+import static org.flickit.assessment.kit.common.ErrorMessageKey.MATURITY_LEVEL_ID_NOT_FOUND;
 
 @Component
 @RequiredArgsConstructor
@@ -26,36 +33,47 @@ public class MaturityLevelPersistenceJpaAdapter implements
     CreateMaturityLevelPort,
     DeleteMaturityLevelPort,
     UpdateMaturityLevelPort,
-    LoadMaturityLevelsPort,
-    LoadAttributeMaturityLevelsPort {
+    LoadAttributeMaturityLevelsPort,
+    LoadMaturityLevelsPort {
 
     private final MaturityLevelJpaRepository repository;
     private final LevelCompetenceJpaRepository levelCompetenceRepository;
+    private final KitDbSequenceGenerators sequenceGenerators;
 
     @Override
     public Long persist(MaturityLevel level, Long kitVersionId, UUID createdBy) {
-        return repository.save(mapToJpaEntityToPersist(level, kitVersionId, createdBy)).getId();
+        var entity = mapToJpaEntityToPersist(level, kitVersionId, createdBy);
+        entity.setId(sequenceGenerators.generateMaturityLevelId());
+        return repository.save(entity).getId();
     }
 
     @Override
     public void delete(Long id, Long kitVersionId) {
+        if (!repository.existsByIdAndKitVersionId(id, kitVersionId))
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
         repository.deleteByIdAndKitVersionId(id, kitVersionId);
     }
 
     @Override
-    public void update(List<MaturityLevel> maturityLevels, Long kitVersionId, UUID lastModifiedBy) {
+    public void updateAll(List<MaturityLevel> maturityLevels, Long kitVersionId, UUID lastModifiedBy) {
         Map<EntityId, MaturityLevel> idToModel = maturityLevels.stream()
-            .collect(Collectors.toMap(
+            .collect(toMap(
                 ml -> new EntityId(ml.getId(), kitVersionId),
                 ml -> ml
             ));
-        List<MaturityLevelJpaEntity> entities = repository.findAllById(idToModel.keySet());
+
+        Set<Long> ids = maturityLevels.stream()
+            .map(MaturityLevel::getId)
+            .collect(Collectors.toSet());
+
+        List<MaturityLevelJpaEntity> entities = repository.findAllByIdInAndKitVersionId(ids, kitVersionId);
         entities.forEach(x -> {
             MaturityLevel newLevel = idToModel.get(new EntityId(x.getId(), kitVersionId));
             x.setIndex(newLevel.getIndex());
             x.setTitle(newLevel.getTitle());
             x.setValue(newLevel.getValue());
             x.setDescription(newLevel.getDescription());
+            x.setTranslations(JsonUtils.toJson(newLevel.getTranslations()));
             x.setLastModificationTime(LocalDateTime.now());
             x.setLastModifiedBy(lastModifiedBy);
         });
@@ -64,30 +82,58 @@ public class MaturityLevelPersistenceJpaAdapter implements
     }
 
     @Override
-    public List<MaturityLevel> loadByKitVersionId(Long kitVersionId) {
-        List<MaturityLevelJpaEntity> maturityLevelEntities = repository.findAllByKitVersionIdOrderByIndex(kitVersionId);
+    public void update(MaturityLevel maturityLevel, Long kitVersionId, LocalDateTime lastModificationTime, UUID lastModifiedBy) {
+        if (!repository.existsByIdAndKitVersionId(maturityLevel.getId(), kitVersionId))
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
 
-        List<Long> levelIds = maturityLevelEntities.stream()
+        repository.update(maturityLevel.getId(),
+            kitVersionId,
+            maturityLevel.getCode(),
+            maturityLevel.getIndex(),
+            maturityLevel.getTitle(),
+            maturityLevel.getDescription(),
+            maturityLevel.getValue(),
+            JsonUtils.toJson(maturityLevel.getTranslations()),
+            lastModificationTime,
+            lastModifiedBy);
+    }
+
+    @Override
+    public void updateOrders(UpdateOrderParam param) {
+        Map<EntityId, UpdateOrderParam.MaturityLevelOrder> idToIndex = param.orders().stream()
+            .collect(Collectors.toMap(
+                ml -> new EntityId(ml.maturityLevelId(), param.kitVersionId()),
+                ml -> ml
+            ));
+
+        Set<Long> ids = param.orders().stream()
+            .map(UpdateOrderParam.MaturityLevelOrder::maturityLevelId)
+            .collect(Collectors.toSet());
+
+        List<MaturityLevelJpaEntity> entities = repository.findAllByIdInAndKitVersionId(ids, param.kitVersionId());
+        if (entities.size() != param.orders().size())
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
+
+        entities.forEach(x -> {
+            UpdateOrderParam.MaturityLevelOrder newLevel = idToIndex.get(new EntityId(x.getId(), param.kitVersionId()));
+            x.setIndex(newLevel.index());
+            x.setValue(newLevel.value());
+            x.setLastModificationTime(param.lastModificationTime());
+            x.setLastModifiedBy(param.lastModifiedBy());
+        });
+        repository.saveAll(entities);
+    }
+
+    @Override
+    public List<MaturityLevel> loadAllByKitVersionId(Long kitVersionId) {
+        var maturityLevelEntities = repository.findAllByKitVersionIdOrderByIndex(kitVersionId);
+
+        var levelIds = maturityLevelEntities.stream()
             .map(MaturityLevelJpaEntity::getId)
             .toList();
 
-        List<LevelCompetenceJpaEntity> levelCompetenceEntities = levelCompetenceRepository.findAllByAffectedLevelIdInAndKitVersionId(levelIds, kitVersionId);
-        Map<Long, List<LevelCompetenceJpaEntity>> levelIdToLevelCompetences = levelCompetenceEntities.stream()
-            .collect(Collectors.groupingBy(LevelCompetenceJpaEntity::getAffectedLevelId));
-
-        return maturityLevelEntities.stream()
-            .map(e -> {
-                MaturityLevel maturityLevel = MaturityLevelMapper.mapToDomainModel(e);
-                if (levelIdToLevelCompetences.containsKey(e.getId())) {
-                    List<LevelCompetenceJpaEntity> levelCompetenceJpaEntities = levelIdToLevelCompetences.get(maturityLevel.getId());
-                    List<MaturityLevelCompetence> maturityLevelCompetences = levelCompetenceJpaEntities.stream()
-                        .map(x -> new MaturityLevelCompetence(x.getEffectiveLevelId(), x.getValue()))
-                        .toList();
-                    maturityLevel.setCompetences(maturityLevelCompetences);
-                } else
-                    maturityLevel.setCompetences(new ArrayList<>());
-                return maturityLevel;
-            }).toList();
+        var levelCompetenceEntities = levelCompetenceRepository.findAllByAffectedLevelIdInAndKitVersionId(levelIds, kitVersionId);
+        return mapToDomainModel(maturityLevelEntities, levelCompetenceEntities);
     }
 
     @Override
@@ -95,5 +141,25 @@ public class MaturityLevelPersistenceJpaAdapter implements
         return repository.loadAttributeLevels(attributeId, kitVersionId).stream()
             .map(e -> new LoadAttributeMaturityLevelsPort.Result(e.getId(), e.getTitle(), e.getIndex(), e.getQuestionCount()))
             .toList();
+    }
+
+    @Override
+    public PaginatedResponse<MaturityLevel> loadByKitVersionId(long kitVersionId, Integer size, Integer page) {
+        String sort = MaturityLevelJpaEntity.Fields.index;
+        Sort.Direction sortDirection = Sort.Direction.ASC;
+
+        var pageResult = repository.findByKitVersionId(kitVersionId,
+            PageRequest.of(page, size, sortDirection, sort));
+
+        var maturityLevels = pageResult.getContent().stream()
+            .map(MaturityLevelMapper::mapToDomainModel)
+            .toList();
+
+        return new PaginatedResponse<>(maturityLevels,
+            pageResult.getNumber(),
+            pageResult.getSize(),
+            sort,
+            sortDirection.name().toLowerCase(),
+            (int) pageResult.getTotalElements());
     }
 }
